requirements.txt
alembic==1.6.5
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
bidict==0.22.0
billiard==3.6.4.0
blinker==1.4
boto3==1.21.39
botocore==1.24.39
bs4==0.0.1
celery==5.2.6
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.14
configparser==5.2.0
controller==0.1.0
croniter==1.3.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
dnspython==2.1.0
docopt==0.6.2
dominate==2.6.0
elastic-transport==7.11.0
elasticsearch==7.9.0
email-validator==1.1.3
emailutils==0.0.1
envparse==0.2.0
et-xmlfile==1.1.0
exotel==0.1.5
f==0.0.1
Faker==15.3.1
filelock==3.7.1
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
flask-core==2.9.0
Flask-HTTPAuth==4.7.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
flask-paginate==2022.1.8
Flask-PostgreSQL==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
guess-language-spirit==0.5.3
gunicorn==20.1.0
h11==0.13.0
httpie==3.2.1
httpresponses==1.0.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.1
jira==3.1.1
jmespath==1.0.0
jsonify==0.5
jsonschema==4.4.0
keyring==23.5.0
kombu==5.2.4
langdetect==1.0.9
Mako==1.1.4
MarkupSafe==2.0.1
mock==4.0.3
MouseInfo==0.1.3
multidict==6.0.2
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-utf-8==0.0.5
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
phantomjs==1.4.1
platformdirs==2.5.2
prison==0.2.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
PyInputPlus==0.2.12
PyJWT==2.1.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySimpleValidate==0.2.12
PySocks==1.7.1
PyStaticConfiguration==0.10.5
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-engineio==4.3.1
python-magic==0.4.25
python-socketio==5.5.2
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.0
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resumable-urlretrieve==0.1.6
rich==12.4.4
rq==1.10.1
s3transfer==0.5.2
serializers==0.2.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sql==2022.4.0
SQLAlchemy==1.4.20
stdiomask==0.0.6
stomp.py==8.0.0
sudo==1.0.0
texttable==1.6.4
tornado==6.2
transliterate==1.10.2
trio==0.21.0
trio-websocket==0.9.2
twilio==6.0.0
tzdata==2022.1
tzlocal==4.2
urllib2-file==0.2.1
urllib3==1.26.6
utils==1.0.1
vine==5.0.0
virtualenv==20.16.2
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.14.1
wsproto==1.1.0
WTForms==2.3.3
zipp==3.8.0
